[scheduler] Use a hook instead of a selector for eventsToRenderGroupedByDay
