+++ packages/x-scheduler/src/material/internals/components/header-toolbar/HeaderToolbar.tsx
@@ -5,7 +5,7 @@ import { useStore } from '@base-ui-components/utils/store';
 import { HeaderToolbarProps } from './HeaderToolbar.types';
 import { ViewSwitcher } from './view-switcher';
 import { useTranslations } from '../../utils/TranslationsContext';
-import { useEventCalendarContext } from '../../hooks/useEventCalendarContext';
+import { useEventCalendarContext } from '../../../../primitives/utils/useEventCalendarContext';
 import { selectors } from '../../../../primitives/use-event-calendar';
 import { SettingsMenu } from './settings-menu';
 import './HeaderToolbar.css';
