+++ packages/x-scheduler/src/material/internals/components/resource-legend/ResourceLegend.tsx
@@ -9,7 +9,7 @@ import { useEventCallback } from '@base-ui-components/utils/useEventCallback';
 import { ResourceLegendProps } from './ResourceLegend.types';
 import { useTranslations } from '../../utils/TranslationsContext';
 import { getColorClassName } from '../../utils/color-utils';
-import { useEventCalendarContext } from '../../hooks/useEventCalendarContext';
+import { useEventCalendarContext } from '../../../../primitives/utils/useEventCalendarContext';
 import { selectors } from '../../../../primitives/use-event-calendar';
 import { CalendarResource } from '../../../../primitives/models';
 import './ResourceLegend.css';
