+++ packages/x-scheduler/src/primitives/use-day-grid-row-event-occurrences/index.ts
@@ -1,0 +1,1 @@
+export * from './useDayGridRowEventOccurrences';
